First shot at the dcpspi daemon.
As it turned out, the dcpspi daemon should really be a kernel driver.
Using spidev and sysfs-GPIO is truly awkward for this protocol, plus
there is potential for race-conditions due to missing bus arbitration
(would require an extra communication channel with the upstream
daemon).

So, this program should be considered a prototype for testing with the
attached slave, but not production-ready code.