Fix spi_read_buffer() implementation.
There were several problems with the implementation concerning correct
buffering, correct unescaping, and doing the actual SPI transfer. We
now use a small buffer for reading data and filter it in-place. It is
appended to the caller's buffer whenever and as much as is possible.
While reading, we are explicitly writing 0xff bytes to the SPI slave
since the kernel driver defaults to sending zeros.