SPI slave indicates ready state by sending 0.
Previously, the SPI slave was assumed to be ready when it sent a
non-NOP byte while polling its ready state (the situation where we are
repetitively sending two NOPs and receiving two bytes).

Now the slave is assumed to be ready to receive data by sending a 0
while polling; any non-zero byte means "do not send" now.

Change-Id: I35752023bdc1395ae0655122a30abadd456f4161