Handle new slave requests during an ongoing transaction.
This situation occurs when we are too slow (maybe due to an unlucky
schedule on a busy system) or if the slave generally applies overly
optimistic assumptions about our response times. In this case, a
timeout on the slave is exceeded, causing it to repeat its request. We
handle this situation by dropping the active transaction and
forgetting the answer we were about to send.

Change-Id: I6ab185050b6adf1be1bbf90411b6d8ebb573ed10