Fix handling of escape sequences in collisions.
If there is a collision because the slave starts sending non-zero,
non-NOP data, and the only non-NOP byte the slave sent was an escape
character after a slide of NOPs, then that escape character got lost.

A unit test has been added to demonstrate the situation in which
triggered the bug.

Change-Id: I62e70491634dad746e2cf7c9f14bfb62e21adcb7