Do not detect SPI collisions by GPIO before polling.
NOTE: This commit temporarily breaks unit tests.

We have a safe fallback already, so there is no point to have the
extra GPIO check. It isn't even faster to check the GPIO because there
are several system calls involved to find out the GPIO state (because
of sysfs interface), and polling must be done anyway.

Change-Id: Id3099eb3abbd6d075be99e42460e4dc518261b80