Fix handling of lost slave requests during collisions (#396).
As a result of details in timing, it is possible for the slave to send
DCP packets in a single SPI read, legally so and without any protocol
violations. Such cases are detected as missed transactions, and this
commit handles them by checking the SPI input buffer for any lost
packets.

Change-Id: I6ce8adad6cc1c90602f362f1cbdaf03ba5fc392a