Dump SPI traffic without change of verbosity level.
There are two extra signals now, RTMIN+7 and RTMIN+8. These can be
used to control dumping of SPI traffic to the log independently of the
global verbosity level.

Signal RTMIN+8 is used to enable SPI traffic dumping. Note that this
setting is not affected by any of the other signals for controlling
verbosity. In particular, RTMIN+0 does not disable traffic dumping if
enabled by RTMIN+8.

Signal RTMIN+7 is used to reset SPI traffic dumping to default
verbosity levels.

Change-Id: I463800429f940de5f70610261c22e460c3438724