Better logs when slave asserts the request pin for a long time.
The message "Ignoring data from DCPD until slave deasserts request
pin" has frequently caused confusion. It looked too much like an error
and didn't quite tell what was actually going on this situation.

This commit changes the message to "Waiting for slave to deassert the
request pin", because this is the intention in this case. There is no
blocking in this case, and requests from DCPD are indeed ignored as
was written in the previous version of this log, but what the software
is doing, logically, is to wait for the request pin being deasserted.

When the pin is deasserted, then a new log message is emitted: "Slave
has deasserted the request pin". It is only emitted if the other
message described above has been emitted as well, so these messages
should always show up in pairs.

Change-Id: I4cd7113e279df3f972cc4e0e280fe469d7e246a2