Add two tests showing the behavior on slight protocol violations.
In case the SPI slave is releasing the request pin while it is waiting
for a pending transaction to be processed by us (two consecutive
messages in very quick succession), but then "loses interest" in
having it processed and thus deasserts its request pin, possibly due
to a timeout on slave side. This would be a protocol violation, and
its outcome depends on the implementation on slave side.

Two assumed behaviors on slave side are simulated in two new unit
tests, and correct recovery on our side is proven by these.

Change-Id: I091f05b335b7a5e186bddaac1a7f6f105e40e1ef